apply principal labels to alert rules defined by grafana-agent (#239)
* apply principal labels to alert rules defined by grafana-agent

* make sure alert rules are present in the k8s charm

* concatenate lists rather than append

+++ src/grafana_agent.py
@@ -110,7 +110,7 @@ class GrafanaAgentCharm(CharmBase):
             dest=charm_root.joinpath(*DASHBOARDS_DEST_PATH.split("/")),
         )
 
-        for rules in [self.loki_rules_paths, self.metrics_rules_paths, self.dashboard_paths]:
+        for rules in [self.loki_rules_paths, self.dashboard_paths]:
             if not os.path.isdir(rules.dest):
                 shutil.copytree(rules.src, rules.dest, dirs_exist_ok=True)
 
@@ -219,6 +219,11 @@ class GrafanaAgentCharm(CharmBase):
         self._update_config()
 
     # Abstract Methods
+    @property
+    def is_k8s(self) -> bool:
+        """Is this a k8s charm."""
+        raise NotImplementedError("Please override the is_k8s method")
+
     def agent_version_output(self) -> str:
         """Gets the raw output from `agent -version`."""
         raise NotImplementedError("Please override the agent_version_output method")
@@ -307,6 +312,7 @@ class GrafanaAgentCharm(CharmBase):
             alerts_func=self.metrics_rules,
             reload_func=self._remote_write.reload_alerts,
             mapping=self.metrics_rules_paths,
+            copy_files=self.is_k8s,  # TODO: This is ugly
         )
 
     def _update_loki_alerts(self):
@@ -314,6 +320,7 @@ class GrafanaAgentCharm(CharmBase):
             alerts_func=self.logs_rules,
             reload_func=self._loki_consumer._reinitialize_alert_rules,
             mapping=self.loki_rules_paths,
+            copy_files=True,
         )
 
     def _update_grafana_dashboards(self):
@@ -329,15 +336,25 @@ class GrafanaAgentCharm(CharmBase):
             return self._recurse_call_chain(maybe_func())
         return maybe_func
 
-    def update_alerts_rules(self, alerts_func: Any, reload_func: Callable, mapping: RulesMapping):
+    def update_alerts_rules(
+        self,
+        alerts_func: Any,
+        reload_func: Callable,
+        mapping: RulesMapping,
+        copy_files: bool = False,
+    ):
         """Copy alert rules from relations and save them to disk."""
         # MetricsEndpointConsumer.alerts is not @property, but Loki is, so
         # do the right thing. With an additional layer of indirection, recurse
         # to the bottom until we find a real List|Dict|not-Callable
         rules = self._recurse_call_chain(alerts_func)
 
-        shutil.rmtree(mapping.dest)
-        shutil.copytree(mapping.src, mapping.dest)
+        if os.path.exists(mapping.dest):
+            shutil.rmtree(mapping.dest)
+        if copy_files:
+            shutil.copytree(mapping.src, mapping.dest)
+        else:
+            os.mkdir(mapping.dest)
         for topology_identifier, rule in rules.items():
             file_handle = pathlib.Path(mapping.dest, "juju_{}.rules".format(topology_identifier))
             file_handle.write_text(yaml.dump(rule))
